// gb-angular-frontend/src/app/interfaces/product-item.ts
export interface ProductItem {
  product_item_id: number;
  item_name: string;
  creation_date: string;
  regular_price: number;
  sale_price: number;
  quantity: number;
  description: string;
  popularity: number;
  status: string;
  product_id: number;
  product_title: string;
  item_image: string;
  item_seq: number;
  action: string;
}